+++ setup.py
@@ -42,4 +42,7 @@ setuptools.setup(
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
     ),
+    extras_require={
+        'idna2008':  ['idna']
+    }
 )
