+++ tox.ini
@@ -4,6 +4,7 @@ envlist = py27,py33,py34,py35,py36,pypy,flake8
 [testenv]
 pip_pre = False
 deps =
+    idna
     -rdev-requirements.txt
 commands =
     py.test {posargs:--cov rfc3986 --cov-fail-under 100 tests/}
